requirements.txt
aiohttp==3.8.5
aiofiles==23.1.0
pyyaml==6.0.1
pyqt5==5.15.9